# src/swagger.yaml
openapi: 3.0.0
info:
  title: NoteMindAI App API
  version: "1.0.0"
  description: API for an AI-powered note-taking application.
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
...
paths:
...
  /categories:
    get:
      summary: Retrieve all categories
      description: Fetches a list of all categories.
      responses:
        '200':
          description: A list of categories.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '500':
          description: Internal server error.
    post:
      summary: Create a new category
      description: Adds a new category to the database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        '200':
          description: Category created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Bad request.
        '500':
          description: Internal server error.
  /categories/{id}:
    get:
      summary: Get a single category by ID
      description: Fetches a category by its ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The category ID.
      responses:
        '200':
          description: A single category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          description: No category found with this ID.
        '500':
          description: Internal server error.
    put:
      summary: Update a category by ID
      description: Updates the details of an existing category.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The category ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        '200':
          description: Category updated successfully.
        '404':
          description: No category found with this ID.
        '500':
          description: Internal server error.
    delete:
      summary: Delete a category by ID
      description: Removes a category from the database.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The category ID.
      responses:
        '200':
          description: Category deleted successfully.
        '404':
          description: No category found with this ID.
        '500':
          description: Internal server error.
  /notes:
    get:
      summary: Retrieve all notes
      description: Fetches a list of all notes.
      responses:
        '200':
          description: A list of notes.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Note'
        '500':
          description: Internal server error.
    post:
      summary: Create a new note
      description: Adds a new note to the database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Note'
      responses:
        '200':
          description: Note created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '400':
          description: Bad request.
        '500':
          description: Internal server error.
  /notes/{id}:
    get:
      summary: Get a single note by ID
      description: Fetches a note by its ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The note ID.
      responses:
        '200':
          description: A single note.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '404':
          description: No note found with this ID.
        '500':
          description: Internal server error.
    put:
      summary: Update a note by ID
      description: Updates the details of an existing note.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The note ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Note'
      responses:
        '200':
          description: Note updated successfully.
        '404':
          description: No note found with this ID.
        '500':
          description: Internal server error.
    delete:
      summary: Delete a note by ID
      description: Removes a note from the database.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The note ID.
      responses:
        '200':
          description: Note deleted successfully.
        '404':
          description: No note found with this ID.
        '500':
          description: Internal server error.
  /sharedNotes:
    get:
      summary: Retrieve all shared notes
      description: Fetches a list of all shared notes, optionally including related Note and User models.
      responses:
        '200':
          description: A list of shared notes.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SharedNote'
        '500':
          description: Internal server error.
    post:
      summary: Create a new shared note relationship
      description: Adds a new shared note relationship to the database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SharedNote'
      responses:
        '200':
          description: Shared note relationship created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedNote'
        '400':
          description: Bad request.
        '500':
          description: Internal server error.
  /sharedNotes/{id}:
    get:
      summary: Get a single shared note by ID
      description: Fetches a shared note by its unique ID (or by a combination of noteId and userId).
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The shared note ID.
      responses:
        '200':
          description: A single shared note.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedNote'
        '404':
          description: No shared note found with this ID.
        '500':
          description: Internal server error.
    put:
      summary: Update shared note permissions
      description: Updates the permissions of an existing shared note relationship.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The shared note ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SharedNote'
      responses:
        '200':
          description: Shared note updated successfully.
        '404':
          description: No shared note found with this ID.
        '500':
          description: Internal server error.
    delete:
      summary: Delete a shared note relationship by ID
      description: Removes a shared note relationship from the database by its unique ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The shared note ID.
      responses:
        '200':
          description: Shared note relationship deleted successfully.
        '404':
          description: No shared note found with this ID.
        '500':
          description: Internal server error.
  /actionableItems:
    get:
      summary: Retrieve all actionable items
      tags: [ActionableItems]
      responses:
        '200':
          description: A list of actionable items.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActionableItem'
        '500':
          description: Internal server error
    post:
      summary: Create a new actionable item
      tags: [ActionableItems]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionableItem'
      responses:
        '200':
          description: Actionable item created successfully.
        '400':
          description: Bad request
        '500':
          description: Internal server error
  /actionableItems/{id}:
    get:
      summary: Find a single actionable item by ID
      tags: [ActionableItems]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID of the actionable item to retrieve
      responses:
        '200':
          description: Actionable item details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionableItem'
        '404':
          description: Actionable item not found
        '500':
          description: Internal server error
    put:
      summary: Update an actionable item's details by ID
      tags: [ActionableItems]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID of the actionable item to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionableItem'
      responses:
        '200':
          description: Actionable item updated successfully.
        '404':
          description: Actionable item not found
        '500':
          description: Internal server error
    delete:
      summary: Remove an actionable item by ID
      tags: [ActionableItems]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID of the actionable item to delete
      responses:
        '200':
          description: Actionable item deleted successfully.
        '404':
          description: Actionable item not found
        '500':
          description: Internal server error
...
components:
  schemas:
...
    Category:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          description: The category ID.
        name:
          type: string
          description: The name of the category.
      example:
        id: 1
        name: "Personal"
    Note:
      type: object
      required:
        - title
        - content
      properties:
        id:
          type: integer
          description: The note ID.
        title:
          type: string
          description: The title of the note.
        content:
          type: string
          description: The content of the note.
      example:
        id: 1
        title: "Meeting Notes"
        content: "Here are the notes from our last meeting..."
    SharedNote:
      type: object
      required:
        - noteId
        - userId
        - permissionType
      properties:
        id:
          type: integer
          description: The unique identifier for the shared note relationship.
        noteId:
          type: integer
          description: The ID of the note being shared.
        userId:
          type: integer
          description: The ID of the user with whom the note is shared.
        permissionType:
          type: string
          enum: [viewer, editor]
          description: The type of permission granted to the shared user.
      example:
        id: 1
        noteId: 10
        userId: 5
        permissionType: editor
    ActionableItem:
      type: object
      properties:
        id:
          type: integer
          description: The ID of the actionable item
        description:
          type: string
          description: The description of the actionable item
        dueDate:
          type: string
          format: date-time
          description: The due date for the actionable item
        status:
          type: string
          enum: [pending, completed]
          description: The status of the actionable item
      required:
        - description
        - status
...